requirements.txt
Django~=1.8.6
djangorestframework~=3.3.1
django-cors-headers~=1.1.0
PyYAML~=3.11
cx_Oracle~=5.2.1
jsonfield~=1.0.3
